performers/python/requirements.txt
couchbase==3.2.7
grpcio==1.44.0
grpcio-tools==1.44.0
protobuf==3.19.4
psutil==5.9.0
